refactor: extract error or to global usings
