+++ Chapter-4-applying-tactical-domain-driven-design/Fitnet.Contracts/Src/Fitnet.Contracts.Api/ContractsEndpoints.cs
@@ -1,5 +1,6 @@
 namespace EvolutionaryArchitecture.Fitnet.Contracts.Api;
 
+using AttachAnnexToBindingContract;
 using Microsoft.AspNetCore.Routing;
 using PrepareContract;
 using SignContract;
