+++ Chapter-4-applying-tactical-domain-driven-design/Fitnet.Contracts/Src/Fitnet.Contracts.Api/GlobalUsings.cs
@@ -1,0 +1,1 @@
+﻿global using ErrorOr;
