+++ Chapter-4-applying-tactical-domain-driven-design/Fitnet.Contracts/Src/Fitnet.Contracts.Api/TerminateBindingContract/TerminateBindingContractEndpoint.cs
@@ -2,7 +2,6 @@ namespace EvolutionaryArchitecture.Fitnet.Contracts.Api.TerminateBindingContract
 
 using Application;
 using Common.Errors;
-using ErrorOr;
 using EvolutionaryArchitecture.Fitnet.Contracts.Application.TerminateBindingContract;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
