+++ Chapter-4-applying-tactical-domain-driven-design/Fitnet.Contracts/Src/Fitnet.Contracts.Core.UnitTests/Common/Assertions/ErrorOr/ErrorOrExtensions.cs
@@ -1,6 +1,4 @@
-﻿namespace EvolutionaryArchitecture.Fitnet.Contracts.Core.UnitTests.Common.Assertions;
-
-using ErrorOr;
+﻿namespace EvolutionaryArchitecture.Fitnet.Contracts.Core.UnitTests.Common.Assertions.ErrorOr;
 
 internal static class ErrorOrExtensions
 {
