+++ Chapter-4-applying-tactical-domain-driven-design/Fitnet.Contracts/Src/Fitnet.Contracts.Core.UnitTests/SignContract/BusinessRules/ContractCanOnlyBeSignedWithin30DaysFromPreparationRuleTests.cs
@@ -1,6 +1,6 @@
 namespace EvolutionaryArchitecture.Fitnet.Contracts.Core.UnitTests.SignContract.BusinessRules;
 
-using Common.Assertions;
+using Common.Assertions.ErrorOr;
 using Core.Common.BussinessRules;
 using Core.SignContract.BusinessRules;
